+++ javascripts/core/automation/turbine_automation.js
@@ -1,0 +1,10 @@
+class DynamoCoilAutomation extends GenericAutomation {
+	constructor(action) {
+		super(action);
+		this.preset = [];
+	}
+
+	selectPreset(preset) {
+		this.preset = preset;
+	}
+}