+++ javascripts/generics/GenericAutomation.js
@@ -1,0 +1,24 @@
+class GenericAutomation {
+	constructor() {
+		this.active = false;
+	}
+
+	toggle() {
+		this.active = !this.active;
+	}
+
+	automate() {}
+}
+
+class GenericActionAutomation {
+	constructor(action) {
+		this.active = false;
+		this.action = action;
+	}
+
+	toggle() {
+		this.active = !this.active;
+	}
+
+	automate() {}
+}